backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
python-dotenv
cloudinary
bcrypt
python-jose[cryptography]